+++ test/Xunit.OpenCategories.UnitTests.V3/ServicesAttributeTests.cs
@@ -5,7 +5,7 @@ namespace Xunit.OpenCategories.UnitTests.V3;
 public class ServicesAttributeTests
 {
     [Fact]
-    public void ServicesWithNoName()
+    public void ServicesWithNoName_ShouldThrowArgumentException()
     {
         var act = () => new ServicesAttribute();
 
@@ -23,7 +23,8 @@ public class ServicesAttributeTests
         testMethod.Should()
             .BeDecoratedWith<FactAttribute>()
             .And.BeDecoratedWith<ServicesAttribute>()
-            .Which.ServiceNames.Should().Contain("Service A")
+            .Which.ServiceNames.Should()
+            .Contain("Service A")
             .And.HaveCount(1);
     }
 
@@ -35,7 +36,8 @@ public class ServicesAttributeTests
         testMethod.Should()
             .BeDecoratedWith<FactAttribute>()
             .And.BeDecoratedWith<ServicesAttribute>()
-            .Which.ServiceNames.Should().Contain("Service A")
+            .Which.ServiceNames.Should()
+            .Contain("Service A")
             .And.Contain("Service B")
             .And.HaveCount(2);
     }
@@ -48,9 +50,60 @@ public class ServicesAttributeTests
         testMethod.Should()
             .BeDecoratedWith<FactAttribute>()
             .And.BeDecoratedWith<ServicesAttribute>()
-            .Which.ServiceNames.Should().Contain("Service A")
+            .Which.ServiceNames.Should()
+            .Contain("Service A")
             .And.Contain("Service B")
             .And.Contain("Service C")
             .And.HaveCount(3);
+    }
+
+    [Fact]
+    public void WhenServicesAreProvided_ThenReturnsServices()
+    {
+        // arrange
+        var attribute = new ServicesAttribute("Service A", "Service B");
+
+        // act
+        var traits = attribute.GetTraits();
+
+        // assert
+        traits.Where(kv => kv.Key.Equals("Service") && kv.Value.Equals("Service A"))
+            .Should()
+            .ContainSingle();
+        traits.Where(kv => kv.Key.Equals("Service") && kv.Value.Equals("Service B"))
+            .Should()
+            .ContainSingle();
+    }
+
+    [Theory]
+    [InlineData("   ")]
+    [InlineData("")]
+    public void WhenServiceNameIsEmptyOrWhitespace_ThenReturnNoServices(string name)
+    {
+        // arrange
+        var attribute = new ServicesAttribute(name);
+
+        // act
+        var traits = attribute.GetTraits();
+
+        // assert
+        traits.Should()
+            .NotContain(kv => kv.Key == "Service");
+    }
+
+    [Theory]
+    [InlineData("Service A", "Service B")]
+    [InlineData("Service A", "Service B, Service C")]
+    public void RegardlessOfServices_ThenShouldHaveCategoryOfService(params string[] names)
+    {
+        // arrange
+        var attribute = new ServicesAttribute(names);
+
+        // act
+        var traits = attribute.GetTraits();
+
+        // assert
+        traits.Should()
+            .Contain(new KeyValuePair<string, string>("Category", "Service"));
     }
 }